polish sidebar and move over tgph components into ui folder
