+++ components/ui/ApiReference/SchemaProperties/helpers.ts
